+++ eva/src/js/componentsSettings.js
@@ -212,7 +212,13 @@ export default {
       'defaultSourceDataField',
       'defaultSourceDataUpdates',
     ],
-    picker: [],
+    picker: [
+      'canChooseTime',
+      'canLastTimeEnter',
+      'canRangeEnter',
+      'canManuallyEnter',
+      'timeOutputFormat',
+    ],
     graph: ['boxShadow'],
     single: [
       'subnumber',
@@ -255,7 +261,7 @@ export default {
     csvg: ['tooltip', 'backgroundImage'],
     ygraph: [],
     bush: [],
-    map: ['osmserver', 'primitivesLibrary'],
+    map: ['osmserver', 'primitives', 'primitivesLibrary'],
     heatmap: [
       'dataFormat',
       'x',
@@ -294,7 +300,7 @@ export default {
       'metricGroup',
       'scatterPlotLegend',
     ],
-    constructorSchemes: ['visible', 'level', 'pinned'],
+    constructorSchemes: ['visible', 'level', 'pinned', 'primitives', 'primitivesLibrary'],
   },
   optionFields: [
     // описание типов полей и их характеристик
@@ -487,6 +493,39 @@ export default {
       elem: 'text-field',
     },
 
+    // datepicker
+    {
+      option: 'canChooseTime',
+      description: 'Выбор времени',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'canLastTimeEnter',
+      description: 'Показать ввод: последние часы/мин/сек',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'canRangeEnter',
+      description: 'Показать выбор диапазона дат',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'canManuallyEnter',
+      description: 'Показать ручной ввод дат',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'timeOutputFormat',
+      description: 'Формат даты для результата',
+      elem: 'text-field',
+      default: '',
+      placeholder: 'Пример: YYYY-MM-DD hh:mm',
+    },
+
     // dashSingle
     {
       option: 'subnumber',
@@ -629,7 +668,6 @@ export default {
         return this.$store.state[this.idDash][this.element]?.availableTableTitles;
       },
     },
-
     {
       option: 'columnCount',
       description: 'Количество столбцов',
@@ -788,6 +826,18 @@ export default {
       elem: 'switch',
     },
 
+    // fullWidthGroup
+    {
+      group: 'Библиотека примитивов',
+      option: 'primitives',
+    },
+    {
+      optionGroup: 'primitives',
+      option: 'primitivesLibrary',
+      elem: 'code-editor',
+      isFullWidth: true,
+    },
+
   ],
   reporstElements: [
     'table',
@@ -869,6 +919,7 @@ export default {
     fromDataSearches: [
       'menu',
       'picker',
+      'constructorSchemes',
     ],
   },
 };