+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- DataPicker: Added settings for displaying the input date and time
+- DataPicker: Added the ability to format the result
+- Added code editor for primitivesLibrary field
+- ConstructorSchemes: Added loading svg icons from server
 - Added the ability to assign an icon via metadata
 - Added ability to correct color via metadata